requirements.txt
async-timeout==4.0.3
datasets==3.1.0
h5py==3.11.0
hf-xet==1.1.10
huggingface-hub==0.24.6
hydra-core==1.3.2
joblib==1.4.2
matplotlib==3.7.5
multidict==6.1.0
multiprocess==0.70.16
numpy==1.24.4
omegaconf==2.3.0
opencv-python==4.10.0.84
pandas==2.0.3
pyarrow==17.0.0
pyparsing==3.1.0
requests==2.32.3
safetensors==0.4.4
scikit-learn==1.3.2
scipy==1.10.1
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
timm==1.0.9
tokenizers==0.19.1
torch==2.4.1
torchvision==0.19.1
triton==3.0.0
tqdm==4.66.5
transformers==4.45.2
